moved the queryArrayParser function to a src/utils/, implemented Mocha
